apify/richard/storelocator/foodlion_com/requirements.txt
requests
sgzipsglogging
